Fix script file
